requirements.txt
mysql-connector-python
yfinance
MetaTrader5
schedule
loguru
retrying
apscheduler
peewee
pymysql